style/improved styles and files restructuring
