+++ frontend/src/App.vue
@@ -1,13 +1,7 @@
 <template>
-  <div class="jumbotron">
-    <div class="container">
-      <div class="row">
-        <div class="col-sm-6 offset-sm-3">
-          <router-view></router-view>
-        </div>
-      </div>
-    </div>
-  </div>
+  <main>
+    <router-view></router-view>
+  </main>
 </template>
 
 <script>
